% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addStatsFromCSV}
\alias{addStatsFromCSV}
\title{Read statistics and nulls from CSV file}
\usage{
addStatsFromCSV(y = NULL, infile, estPi0 = FALSE)
}
\arguments{
\item{y}{a SummarizedExperiment (if NULL, function will
output a data.frame)}

\item{infile}{character, path to the \code{summary.csv} file}

\item{estPi0}{logical, see \code{\link{swish}}}
}
\value{
the SummarizedExperiment with metadata columns added,
or if \code{y} is NULL, a data.frame of compiled results
}
\description{
After running \code{\link{splitSwish}} and the associated
\code{Snakefile}, this function can be used to gather and
add the results to the original object. See the alevin
section of the vignette for an example.
}
