% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deswish.R
\name{deswish}
\alias{deswish}
\title{deswish: DESeq2 With Inferential Samples Helps}
\usage{
deswish(y, x, coef)
}
\arguments{
\item{y}{a SummarizedExperiment containing the inferential replicate matrices,
as output by \code{tximeta}}

\item{x}{the design matrix}
}
\value{
a SummarizedExperiment with metadata columns added
}
\description{
The DESeq2 With Inferential Samples implementation supposes
a hierarchical distribution of log2 fold changes.
The final posterior standard deviation is calculated by
adding the posterior variance from modeling biological replicates
(apeglm) and the observed variance on the posterior mode
over inferential replicates.
}
\references{
The \code{DESeq} and \code{lfcShrink} function in the \code{DESeq2} package.

Zhu, Ibrahim, Love "Heavy-tailed prior distributions for sequence count data:
removing the noise and preserving large differences" Bioinformatics (2018).

Love, Huber, Anders "Moderated estimation of fold change and dispersion
for RNA-seq data with DESeq2" Genome Biology (2014).
}
