% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpond.R
\docType{package}
\name{fishpond-package}
\alias{fishpond}
\alias{fishpond-package}
\title{Fishpond: downstream methods and tools for expression data}
\description{
This package provides statistical methods and other tools for
working with Salmon and Alevin quantification of RNA-seq data.
Fishpond contains the Swish non-parametric method for
detecting differential transcript expression (DTE). Swish can
also be used to detect differential gene expresion (DGE),
to perform allelic analysis, or to assess changes in isoform
proportions.
}
\details{
The main Swish functions are:
\itemize{
\item \code{\link{scaleInfReps}} - scaling transcript or gene expression data
\item \code{\link{labelKeep}} - labelling which features have sufficient counts
\item \code{\link{swish}} - perform non-parametric differential analysis
\item Plots, e.g., \code{\link{plotMASwish}}, \code{\link{plotInfReps}}
}

All software-related questions should be posted to the Bioconductor Support Site:

\url{https://support.bioconductor.org}

The code can be viewed at the GitHub repository,
which also lists the contributor code of conduct:

\url{https://github.com/mikelove/fishpond}
}
\references{
Swish method:

Zhu, A., Srivastava, A., Ibrahim, J.G., Patro, R., Love, M.I. (2019)
Nonparametric expression analysis using inferential replicate counts.
Nucleic Acids Research.
\url{https://doi.org/10.1093/nar/gkz622}

Compression, makeInfReps and splitSwish:

Van Buren, S., Sarkar, H., Srivastava, A., Rashid, N.U.,
Patro, R., Love, M.I. (2020)
Compression of quantification uncertainty for scRNA-seq counts.
bioRxiv.
\url{https://doi.org/10.1101/2020.07.06.189639}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thelovelab.github.io/fishpond}
  \item \url{https://thelovelab.com/mikelove/fishpond}
  \item Report bugs at \url{https://support.bioconductor.org/tag/fishpond}
}

}
\author{
\strong{Maintainer}: Michael Love \email{michaelisaiahlove@gmail.com}

Authors:
\itemize{
  \item Anqi Zhu [contributor]
  \item Avi Srivastava [contributor]
  \item Rob Patro [contributor]
  \item Joseph Ibrahim [contributor]
}

Other contributors:
\itemize{
  \item Hirak Sarkar [contributor]
  \item Euphy Wu [contributor]
  \item Noor Pratap Singh [contributor]
  \item Scott Van Buren [contributor]
  \item Dongze He [contributor]
  \item Steve Lianoglou [contributor]
  \item Wes Wilson [contributor]
  \item Jeroen Gilis [contributor]
}

}
\keyword{package}
