% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeInfReps}
\alias{makeInfReps}
\title{Make inferential replicates from mean and variance}
\usage{
makeInfReps(y, numReps)
}
\arguments{
\item{y}{a SummarizedExperiment}

\item{numReps}{how many inferential replicates}
}
\value{
a SummarizedExperiment
}
\description{
Creates simulated inferential replicate counts from
\code{mean} and \code{variance} assays.
}
