% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-allelic.R
\name{plotAllelicGene}
\alias{plotAllelicGene}
\title{Plot allelic counts in a gene context using Gviz}
\usage{
plotAllelicGene(
  y,
  gene,
  db,
  region = NULL,
  genome = NULL,
  tpmFilter = 1,
  countFilter = 10,
  pc = 1,
  transcriptAnnotation = "symbol",
  statCol = "black",
  allelicCol = c("dodgerblue", "goldenrod1"),
  isoformCol = "firebrick",
  bgCol = "grey60",
  ideogram = TRUE
)
}
\arguments{
\item{y}{a SummarizedExperiment (see \code{swish})}

\item{gene}{the name of the gene of interest, requires
a column \code{gene_id} in the metadata columns of the
rowRanges of y}

\item{db}{either a TxDb or EnsDb object}

\item{region}{GRanges, the region to be displayed in the Gviz plot.
if not specified, will be set according to the gene plus 20%
of the total gene extent on either side}

\item{genome}{UCSC genome code (e.g. \code{"hg38"},
if not provided it will use the genome() of the rowRanges of \code{y}}

\item{tpmFilter}{minimum TPM value (mean over samples) to keep a feature}

\item{countFilter}{minimum count value (mean over samples) to keep a feature}

\item{pc}{pseudocount to avoid dividing by zero in allelic proportion}

\item{transcriptAnnotation}{argument passed to Gviz::GeneRegionTrack
(\code{"symbol"}, \code{"gene"}, \code{"transcript"}, etc.)}

\item{statCol}{the color of the lollipops for q-value and log2FC}

\item{allelicCol}{the colors of the lines for allelic proportion}

\item{isoformCol}{the colors of the lines for isoform proportion}

\item{bgCol}{background color of the axis labels}

\item{ideogram}{logical, whether to include ideogram}
}
\value{
nothing, a plot is displayed
}
\description{
Plot allelic data (allelic proportions, isoform propostions)
in a gene context leveraging the Gviz package. See the allelic
vignette for example usage. TPM and count filters are used by
default to clean up the plot of features with minimal signal;
note that the isoform proportion displayed at the bottom of the
plot is among the features that pass the filtering steps.
}
