% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-allelic.R
\name{plotAllelicHeatmap}
\alias{plotAllelicHeatmap}
\title{Plot allelic ratio heatmap}
\usage{
plotAllelicHeatmap(y, idx)
}
\arguments{
\item{y}{a SummarizedExperiment (see \code{swish})}

\item{idx}{a numeric or logical vector of which features
to plot}
}
\value{
nothing, a plot is displayed
}
\description{
Plot allelic ratio heatmap over features and samples
using the pheatmap package.
}
