% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotInfReps}
\alias{plotInfReps}
\title{Plot inferential replicates for a gene or transcript}
\usage{
plotInfReps(y, idx, x, cov = NULL, cols.drk = c("dodgerblue",
  "goldenrod4"), cols.lgt = c("lightblue1", "goldenrod1"),
  xaxis = TRUE)
}
\arguments{
\item{y}{a SummarizedExperiment (see \code{swish})}

\item{idx}{the name or row number of the gene or transcript}

\item{x}{the name of the condition variable}

\item{cov}{the name of the covariate for adjustment}

\item{cols.drk}{dark colors for the lines of the boxes}

\item{cols.lgt}{light colors for the inside of the boxes}

\item{xaxis}{logical, whether to label the sample numbers}
}
\value{
nothing, a plot is displayed
}
\description{
Plot inferential replicates for a gene or transcript
}
\examples{

y <- makeSimSwishData()
plotInfReps(y, 3, "condition")

y <- makeSimSwishData(n=40)
y$batch <- factor(rep(c(1,2,3,1,2,3),c(5,10,5,5,10,5)))
plotInfReps(y, 3, "condition", "batch", xaxis=FALSE)

}
