% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotMASwish}
\alias{plotMASwish}
\title{MA plot}
\usage{
plotMASwish(y, alpha = 0.05, sigcolor = "blue", ...)
}
\arguments{
\item{y}{a SummarizedExperiment (see \code{swish})}

\item{alpha}{the FDR threshold for coloring points}

\item{sigcolor}{the color for the significant points}

\item{...}{passed to plot}
}
\value{
nothing, a plot is displayed
}
\description{
MA plot
}
\examples{

y <- makeSimSwishData()
y <- scaleInfReps(y)
y <- labelKeep(y)
y <- swish(y, x="condition")
plotMASwish(y)

}
