% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevin.R
\name{readEDS}
\alias{readEDS}
\title{readEDS - a utility function for quickly reading in Alevin's EDS format}
\usage{
readEDS(files, num.genes, num.cells)
}
\arguments{
\item{files}{pointer to the EDS file}

\item{num.genes}{number of genes}

\item{num.cells}{number of cells}
}
\value{
a list, the positions and the expression values of
the non-zero expressions in the Alevin EDS file.
\strong{NOTE:} the positions are 0-based, i.e. ranging from
[0, num.cells-1]
}
\description{
readEDS - a utility function for quickly reading in Alevin's EDS format
}
