% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon-EC.R
\name{salmonEC}
\alias{salmonEC}
\title{Construct a matrix of transcript compatibility counts from salmon output}
\usage{
salmonEC(paths, tx2gene, multigene = FALSE, quiet = FALSE)
}
\arguments{
\item{paths}{`Charachter` or `character vector`, full path specifying the 
location of the `eq_classes.txt` files generated with salmon.}

\item{tx2gene}{A `dataframe` linking transcript identifiers to their 
corresponding gene identifiers. Transcript identifiers must be in a column 
`isoform_id`. Corresponding gene identifiers must be in a column `gene_id`.}

\item{multigene}{`Logical`, should equivalence classes that are compatible 
with multiple genes be retained? Default is `FALSE`, removing such ambiguous
equivalence classes.}

\item{quiet}{`Logical`, set `TRUE` to avoid displaying messages.}
}
\value{
a sparse count matrix
}
\description{
Constructs a count matrix with equivalence class identifiers
in the rows and barcode identifiers in the columns. The count matrix is 
generated from one or multiple `eq_classes.txt` files that have been created
by running salmon with the --dumpEq flag. Salmon - 
\url{https://doi.org/10.1038/nmeth.4197}
}
\author{
Jeroen Gilis
}
