% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scaleInfRep}
\alias{scaleInfRep}
\title{Scale inferential replicate counts}
\usage{
scaleInfRep(se, meanDepth = NULL, sfFun = NULL, minCount = 10,
  minN = 3)
}
\arguments{
\item{se}{a SummarizedExperiment with: \code{infReps} a list of
inferential replicate count matrices, \code{counts} the
estimated counts matrix, and \code{length} the effective
lengths matrix}

\item{meanDepth}{(optional) user can
specify a different mean sequencing depth. By default
the geometric mean sequencing depth is computed}

\item{sfFun}{(optional) size factors function. An
alternative to the median ratio can be provided here to adjust
the scaledTPM so as to remove remaining library size differences}

\item{minCount}{for internal filtering, the minimum count}

\item{minN}{for internal filtering, the minimum sample size at \code{minCount}}
}
\value{
a SummarizedExperiment wiht the inferential replicates
as scaledTPM with library size already corrected (no need for further
normalization)
}
\description{
A helper function to scale the inferential replicates
to the mean sequencing depth. The scaling takes into account
a robust estimator of size factor (median ratio method is used).
First, counts are converted to TPM, then scaled to
the mean sequence depth, finally the scaledTPM are
adjusted up or down by the median ratio size factor to
minimize systematic differences across samples.
}
