% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swish.R
\name{swish}
\alias{swish}
\title{swish: SAMseq With Inferential Samples Helps}
\usage{
swish(y, x, cov = NULL, nperms = 30, estPi0 = FALSE,
  wilcoxP = NULL)
}
\arguments{
\item{y}{a SummarizedExperiment containing the inferential replicate
matrices of median-ratio-scaled TPM as assays}

\item{x}{the name of the condition variable}

\item{cov}{the name of the covariate for adjustment.
if provided a stratified Wilcoxon in performed}

\item{nperms}{the number of permutations}

\item{estPi0}{logical, whether to estimate pi0}
}
\value{
a SummarizedExperiment with metadata columns added
}
\description{
swish: SAMseq With Inferential Samples Helps
}
\examples{

y <- makeSwishData()
assayNames(y)
y <- labelKeep(y)
y <- scaleInfReps(y)
y <- swish(y, x="condition")
stat <- mcols(y)$stat
hist(stat,breaks=40,col="grey")
cols = rep(c("blue","purple","red"),each=2)
for (i in 1:6) {
  arrows(stat[i], 20, stat[i], 10, col=cols[i], length=.1, lwd=2)
}

}
\references{
The \code{SAMseq} function in the \code{samr} package.

Jun Li and Robert Tibshirani "Finding consistent patterns:
A nonparametric approach for identifying differential expression
in RNA-Seq data" Stat Methods Med Res (2013).
}
