% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{writeExampleFryDat}
\alias{writeExampleFryDat}
\title{Serializes example fry output data from an `example-fry-dat.csv` as if
it were produced from an alevin-fry quant run}
\usage{
writeExampleFryDat(x = "fry-usa-basic", ...)
}
\arguments{
\item{x}{the name of the top level directory in
`extdata/alevin/example-fry-USA-quants` or a list result from calling
`readExampleFryData`}
}
\description{
Serializes example fry output data from an `example-fry-dat.csv` as if
it were produced from an alevin-fry quant run
}
