% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addStatsFromCSV}
\alias{addStatsFromCSV}
\title{Read statistics and nulls from CSV file}
\usage{
addStatsFromCSV(y = NULL, infile, estPi0 = FALSE)
}
\arguments{
\item{y}{a SummarizedExperiment (if NULL, function will
output a data.frame)}

\item{infile}{character, path to CSV file}

\item{estPi0}{logical, see \code{\link{swish}}}
}
\value{
data.frame of compiled results
}
\description{
Read statistics and nulls from CSV file
}
