% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevin-EC.R
\name{alevinEC}
\alias{alevinEC}
\title{Construct a matrix of transcript compatibility counts from alevin 
output}
\usage{
alevinEC(paths, tx2gene, multigene = FALSE, quiet = FALSE)
}
\arguments{
\item{paths}{`Charachter` or `character vector`, full path specifying the 
location of the `bfh.txt` files generated with alevin-fry.}

\item{tx2gene}{A `dataframe` linking transcript identifiers to their 
corresponding gene identifiers. Transcript identifiers must be in a column 
`isoform_id`. Corresponding gene identifiers must be in a column `gene_id`.}

\item{multigene}{`Logical`, should equivalence classes that are compatible 
with multiple genes be retained? Default is `FALSE`, removing such ambiguous
equivalence classes.}

\item{quiet}{`Logical`, set `TRUE` to avoid displaying messages.}
}
\value{

}
\description{
Constructs a UMI count matrix with equivalence class identifiers
in the rows and barcode identifiers in thecolumn. The count matrix is 
generated from one or multiple `bfh.txt` files that have ben generated by 
alevin-fry when run with the --dumpBFH flag. Alevin-fry - 
\url{https://doi.org/10.1186/s13059-019-1670-y}
}
\details{

}
