% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swish.R
\docType{package}
\name{fishpond-package}
\alias{fishpond-package}
\title{Downstream methods for Salmon and Alevin expression data}
\description{
This package provides statistical methods and other tools for
working with Salmon and Alevin quantification of RNA-seq data.
In particular, it contains the Swish non-parametric method for
detecting differential transcript expression (DTE). Swish can
also be used to detect differential gene expresion (DGE).
}
\details{
The main functions are:
\itemize{
\item \code{\link{scaleInfReps}} - scaling transcript or gene expression data
\item \code{\link{labelKeep}} - labelling which features have sufficient counts
\item \code{\link{swish}} - perform non-parametric differential analysis
\item Plots, e.g., \code{\link{plotMASwish}}, \code{\link{plotInfReps}}
}

All software-related questions should be posted to the Bioconductor Support Site:

\url{https://support.bioconductor.org}

The code can be viewed at the GitHub repository,
which also lists the contributtor code of conduct:

\url{https://github.com/mikelove/fishpond}
}
\references{
Zhu, A., Srivastava, A., Ibrahim, J.G., Patro, R., Love, M.I. (2019)
Nonparametric expression analysis using inferential replicate counts.
Nucleic Acids Research.
\url{https://doi.org/10.1093/nar/gkz622}
}
\author{
Anqi Zhu, Avi Srivastava, Joseph G. Ibrahim, Rob Patro, Michael I. Love
}
\keyword{package}
