% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{importAllelicCounts}
\alias{importAllelicCounts}
\title{Import allelic counts as a SummarizedExperiment}
\usage{
importAllelicCounts(
  coldata,
  a1,
  a2,
  format = c("wide", "assays"),
  tx2gene = NULL,
  ...
)
}
\arguments{
\item{coldata}{a data.frame as used in \code{tximeta}}

\item{a1}{the symbol for the effect/alternative allele}

\item{a2}{the symbol for the reference allele}

\item{format}{either \code{"wide"} or \code{"assays"} for whether
to combine the allelic counts as columns (wide) or put the allelic
count information in different assay slots (assays). For wide output, the
reference allele (a2) comes first, then the alternative allele (a1),
e.g. \code{[a2 | a1]}.
For assays output, all of the original matrices are renamed with a prefix,
either \code{a1-} or \code{a2-}.}

\item{tx2gene}{optional, a data.frame with first column indicating
transcripts, second column indicating genes (or any other transcript
grouping). Either this should include the \code{a1} and \code{a2}
suffix for the transcripts and genes, or those will be added internally,
if it is detected that the first transcript does not have these suffices.
For example if \code{_alt} or \code{_ref}, or \code{_M} or \code{_P}
(as indicated by the \code{a1} and \code{a2} arguments) are not present
in the table, the table rows will be duplicated with those suffices
added on behalf of the user.
If not provided, the output object will be transcript-level.}

\item{...}{any arguments to pass to tximeta}
}
\value{
a SummarizedExperiment, with allele counts (and other data)
combined into a wide matrix \code{[a2 | a1]}, or as assays (a1, then a2).
The original strings associated with a1 and a2 are stored in the
metadata of the object, in the \code{alleles} list element.
Note the reference level of \code{se$allele} will be \code{"a2"}, and
so comparisons by default will be a1/a2 (alt/ref).
}
\description{
Read in Salmon quantification of allelic counts from a
diploid transcriptome. Assumes that diploid transcripts
are marked with the following suffix: an underscore and
a consistent symbol for each of the two alleles,
e.g. \code{ENST123_M} and \code{ENST123_P},
or \code{ENST123_alt} (a1) and \code{ENST123_ref} (a2).
There must be exactly two alleles for each reference transcript, 
and the \code{--keep-duplicates} option should be used in
Salmon indexing to avoid removing transcripts with identical sequence.
The output object has half the number of transcripts,
with the two alleles either stored in a \code{"wide"} object,
or as re-named \code{"assays"}. Note carefully that the symbol
provided to \code{a1} is used as the alternative allele,
and \code{a2} is used as the reference allele
(and therefore \code{"a2"} is the reference level of the
\code{allele} factor that is returned in the colData).
}
\details{
Requires the tximeta package.
\code{skipMeta=TRUE} is used, as it is assumed
the diploid transcriptome does not match any reference
transcript collection. This may change in future iterations
of the function, depending on developments in upstream
software.
}
