% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-allelic.R
\name{plotAllelicGene}
\alias{plotAllelicGene}
\title{Plot allelic counts in a gene context using Gviz}
\usage{
plotAllelicGene(
  y,
  gene,
  db,
  region = NULL,
  genome = NULL,
  tpmFilter = 1,
  countFilter = 10,
  pc = 1,
  transcriptAnnotation = "symbol",
  ideogram = TRUE,
  allelicCol = c("dodgerblue", "goldenrod1"),
  isoformCol = "firebrick",
  statCol = "black",
  gridCol = "grey80",
  baselineCol = "black",
  titleCol = "black",
  titleAxisCol = "black",
  titleBgCol = "white",
  geneBorderCol = "darkblue",
  geneFillCol = "darkblue",
  genomeAxisCol = "black",
  innerFontCol = "black",
  ...
)
}
\arguments{
\item{y}{a SummarizedExperiment (see \code{swish})}

\item{gene}{the name of the gene of interest, requires
a column \code{gene_id} in the metadata columns of the
rowRanges of y}

\item{db}{either a TxDb or EnsDb object to use for the gene model}

\item{region}{GRanges, the region to be displayed in the Gviz plot.
if not specified, will be set according to the gene plus 20%
of the total gene extent on either side}

\item{genome}{UCSC genome code (e.g. \code{"hg38"},
if not specified it will use the \code{GenomeInfoDb::genome()}
of the rowRanges of \code{y}}

\item{tpmFilter}{minimum TPM value (mean over samples) to keep a feature}

\item{countFilter}{minimum count value (mean over samples) to keep a feature}

\item{pc}{pseudocount to avoid dividing by zero in allelic proportion calculation}

\item{transcriptAnnotation}{argument passed to Gviz::GeneRegionTrack
(\code{"symbol"}, \code{"gene"}, \code{"transcript"}, etc.)}

\item{ideogram}{logical, whether to include ideogram}

\item{allelicCol}{the colors of the points and lines for allelic proportion}

\item{isoformCol}{the colors of the points and lines for isoform proportion}

\item{statCol}{the color of the lollipops for q-value and log2FC}

\item{gridCol}{the color of the grid in the data tracks}

\item{baselineCol}{the color of the horizontal baseline for q-value and lo2gFC}

\item{titleCol}{font color of the side titles (track labels)}

\item{titleAxisCol}{axis color of the side titles (track labels)}

\item{titleBgCol}{background color of the side titles (track labels)}

\item{geneBorderCol}{the color of the borders and font in gene region track}

\item{geneFillCol}{the color of the fill in the gene region track}

\item{genomeAxisCol}{line color of the genome axis track}

\item{innerFontCol}{font color of genome axis track, ideogram, and
allelic proportion legend}

\item{...}{additional arguments passed to \code{Gviz::plotTracks()}}
}
\value{
nothing, a plot is displayed
}
\description{
Plot allelic data (allelic proportions, isoform propostions)
in a gene context leveraging the Gviz package. See the allelic
vignette for example usage. TPM and count filters are used by
default to clean up the plot of features with minimal signal;
note that the isoform proportion displayed at the bottom of the
plot is among the features that pass the filtering steps.
}
