% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotInfReps}
\alias{plotInfReps}
\title{Plot inferential replicates for a gene or transcript}
\usage{
plotInfReps(
  y,
  idx,
  x,
  cov = NULL,
  cols.drk = c("dodgerblue", "goldenrod4", "green4", "red3", "purple4", "royalblue4"),
  cols.lgt = c("lightblue1", "goldenrod1", "lightgreen", "salmon1", "orchid1",
    "royalblue1"),
  xaxis,
  xlab,
  ylim,
  legend = FALSE,
  useMean = TRUE,
  applySF = FALSE,
  reorder,
  thin
)
}
\arguments{
\item{y}{a SummarizedExperiment (see \code{swish})}

\item{idx}{the name or row number of the gene or transcript}

\item{x}{the name of the condition variable}

\item{cov}{the name of the covariate for adjustment}

\item{cols.drk}{dark colors for the lines of the boxes}

\item{cols.lgt}{light colors for the inside of the boxes}

\item{xaxis}{logical, whether to label the sample numbers.
default is \code{TRUE} if there are less than 30 samples}

\item{xlab}{the x-axis label}

\item{ylim}{y limits}

\item{legend}{logical, show simple legend (default FALSE)}

\item{useMean}{logical, when inferential replicates
are not present, use the \code{mean} assay or the
\code{counts} assay for plotting}

\item{applySF}{logical, when inferential replicates are
not present, should \code{y$sizeFactor} be divided out
from the mean and interval plots (default FALSE)}

\item{reorder}{logical, should points within a group
defined by condition and covariate be re-ordered by
their count value (default is FALSE, except for alevin data)}

\item{thin}{integer, should the mean and interval lines
be drawn thin (the default switches from 0 [not thin]
to 1 [thinner] at n=150 cells, and from 1 [thinner]
to 2 [thinnest] at n=400 cells)}
}
\value{
nothing, a plot is displayed
}
\description{
For datasets with inferential replicates, boxplots are
drawn for the two groups and potentially grouped by
covariates. For datasets with only mean and variance,
points and intervals (95% intervals using Normal
approximation) are drawn.
}
\examples{

y <- makeSimSwishData()
plotInfReps(y, 3, "condition")

y <- makeSimSwishData(n=40)
y$batch <- factor(rep(c(1,2,3,1,2,3),c(5,10,5,5,10,5)))
plotInfReps(y, 3, "condition", "batch")

}
