% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDS.R
\name{readEDS}
\alias{readEDS}
\title{readEDS - a utility function for quickly reading in Alevin's EDS format}
\usage{
readEDS(numOfGenes, numOfOriginalCells, countMatFilename, tierImport = FALSE)
}
\arguments{
\item{numOfGenes}{number of genes}

\item{numOfOriginalCells}{number of cells}

\item{countMatFilename}{pointer to the EDS file, \code{quants_mat.gz}}

\item{tierImport}{whether the \code{countMatFilename} refers to a quants tier file}
}
\value{
a genes x cells sparse matrix, of the class \code{dgCMatrix}
}
\description{
readEDS - a utility function for quickly reading in Alevin's EDS format
}
