% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swish.R
\name{swish}
\alias{swish}
\title{swish: SAMseq With Inferential Samples Helps}
\usage{
swish(y, x, cov = NULL, nperms = 30, estPi0 = FALSE)
}
\arguments{
\item{y}{a SummarizedExperiment containing the inferential replicate
matrices of median-ratio-scaled TPM as assays}

\item{x}{the name of the condition variable}

\item{cov}{the name of the covariate for adjustment.
if provided a stratified Wilcoxon in performed}

\item{nperms}{the number of permutations}

\item{estPi0}{logical, whether to estimate pi0}
}
\value{
a SummarizedExperiment with metadata columns added
}
\description{
swish: SAMseq With Inferential Samples Helps
}
\references{
The \code{SAMseq} function in the \code{samr} package.

Jun Li and Robert Tibshirani "Finding consistent patterns:
A nonparametric approach for identifying differential expression
in RNA-Seq data" Stat Methods Med Res (2013).
}
