% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swish.R
\name{swish}
\alias{swish}
\title{swish: SAMseq With Inferential Samples Helps}
\usage{
swish(y, x, cov = NULL, pair = NULL, nperms = 30, wilcoxP = 0.25,
  estPi0 = FALSE, qvaluePkg = "samr", pc = 5, quiet = FALSE)
}
\arguments{
\item{y}{a SummarizedExperiment containing the inferential replicate
matrices of median-ratio-scaled TPM as assays 'infRep1', 'infRep2', etc.}

\item{x}{the name of the condition variable. A factor with two
levels for a two group analysis (possible to adjust for covariate
or matched samples, see next two arguments)}

\item{cov}{the name of the covariate for adjustment.
If provided a stratified Wilcoxon in performed.
Cannot be used with \code{pair}}

\item{pair}{the name of the pair variable, which should be the
number of the pair. Can be an integer or factor.
If specified, a signed rank test is used
to build the statistic. All samples across \code{x} must be
pairs if this is specified. Cannot be used with \code{cov}.}

\item{nperms}{the number of permutations}

\item{wilcoxP}{the quantile of the Wilcoxon statistics across
inferential replicates to use as the test statistic.
If set to NULL this will use the mean over inferential replicates}

\item{estPi0}{logical, whether to estimate pi0}

\item{qvaluePkg}{character, which package to use for q-value estimation,
\code{samr} or \code{qvalue}}

\item{pc}{pseudocount for finite estimation of \code{log2FC}, not used
in calculation of test statistics, \code{locfdr} or \code{qvalue}}

\item{quiet}{display no messages}
}
\value{
a SummarizedExperiment with metadata columns added:
the statistic (either a centered Wilcoxon Mann-Whitney
or a signed rank statistic, aggregated over inferential replicates),
a log2 fold change (the median over inferential replicates,
and averaged over pairs or groups (if groups, weighted by sample size),
the local FDR and q-value, as estimated by the \code{samr} package.
}
\description{
swish: SAMseq With Inferential Samples Helps
}
\examples{

library(SummarizedExperiment)
set.seed(1)
y <- makeSimSwishData()
y <- scaleInfReps(y)
y <- labelKeep(y)
y <- swish(y, x="condition")

# histogram of the swish statistics
hist(mcols(y)$stat, breaks=40, col="grey")
cols = rep(c("blue","purple","red"),each=2)
for (i in 1:6) {
  arrows(mcols(y)$stat[i], 20,
         mcols(y)$stat[i], 10,
         col=cols[i], length=.1, lwd=2)
}

# plot inferential replicates
plotInfReps(y, 1, "condition")
plotInfReps(y, 3, "condition")
plotInfReps(y, 5, "condition")

}
\references{
\code{swish} is described in the following reference:

Anqi Zhu, Avi Srivastava, Joseph G Ibrahim, Rob Patro, Michael I Love
"Nonparametric expression analysis using inferential replicate counts"
bioRxiv (2019).

The \code{swish} method builds upon the \code{SAMseq} method,
and extends it by incorporating inferential uncertainty.
\code{swish} internally calls functions from the \code{samr}
package, for example, the calculation of local FDR and q-value.

The citation for \code{SAMseq} is:

Jun Li and Robert Tibshirani "Finding consistent patterns:
A nonparametric approach for identifying differential expression
in RNA-Seq data" Stat Methods Med Res (2013).
}
